refactor(runtime): Message type, handle and call methods.
Although very heartbreaking, this refactor adds a ton in quality of
life for the lib itself. The user will interact with the handler function
for each rpc in a similar way that they'd do so if they were using Go counterpart of this lib.
They'd use serde_value::from_value in a similar way of golang's json.Marshall